Multilingual (#8)
This merge introduces multilingual scenarios support as well as a couple of new features:

Running scripts is now done via scenarios in ./run.sh, which adds a config arg if needed.
Common logic of scripts like train.py and validate.py is moved to ./src/routines package, but their code needs to be refactored properly.
ModelIO and DefaultTrainer were restructured in order to provide schedulers support, however, ModelIO code is far from clean, needs a bit of refactoring.
Scripts which work with mswc dataset on conversion from opus to wav, transforming splits.csv and converting wav to spectrograms were added.
Opus to wav conversion was made a bit faster with multiprocessing.Pool.
FewShotDataloader was deleted as its responsibility was more suitable for corresponding Dataset. Now ClassificationDataset is used both for classification and few-shot as both tasks are classification tasks but with different underlying data.
Code was refactored and commented a bit. However scripts in ./src and ./src/routines still need to be improved for better readability.